2005_Greeninja/requirements.txt
torch>=2.1.0
numpy>=1.23.0
Pillow>=9.0.0
